+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--recreate` flag for `git clone` command that recreates previously cloned repository
+- `status` command
 
 ## [0.3.0] - 2022-07-28